+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## v0.1.1
 
-- Indution macro on nats
+- Exit codes (1 for logical error, 2 for syntax error, 3 for file error)
+
+- "check" statement to run Z3 on a goal with no relation to the program
+
+- Induction macros for nats
 
 - Syntax: axioms before the program